backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy>=2.0.23
openai==1.3.6
qdrant-client
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.2
google-generativeai